# docs/swagger.yaml
definitions:
  schemas.Rol:
    properties:
      id_rol:
        type: integer
      rol:
        type: string
    type: object
  schemas.Usuario:
    properties:
      apellido:
        type: string
      correo:
        type: string
      estado:
        type: boolean
      fecha_aceptacion:
        type: string
      fecha_solicitud:
        type: string
      id_rol:
        type: integer
      id_usuario:
        type: integer
      nombre:
        type: string
      pass:
        type: string
      rol:
        $ref: '#/definitions/schemas.Rol'
    type: object
info:
  contact: {}
paths:
  /api/v1/roles:
    get:
      description: Obtiene un listado completo de roles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schemas.Rol'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar todos los roles
      tags:
      - Roles
    post:
      consumes:
      - application/json
      description: Crea un rol en la base de datos con los datos proporcionados
      parameters:
      - description: Datos del rol
        in: body
        name: rol
        required: true
        schema:
          $ref: '#/definitions/schemas.Rol'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.Rol'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Crear un nuevo rol
      tags:
      - Roles
  /api/v1/roles/{id}:
    delete:
      description: Elimina un rol de la base de datos usando su ID
      parameters:
      - description: ID del rol
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Eliminar rol
      tags:
      - Roles
    get:
      description: Obtiene un rol específico usando su ID
      parameters:
      - description: ID del rol
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.Rol'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Obtener rol por ID
      tags:
      - Roles
    put:
      consumes:
      - application/json
      description: Actualiza los datos de un rol existente
      parameters:
      - description: ID del rol
        in: path
        name: id
        required: true
        type: integer
      - description: Datos actualizados del rol
        in: body
        name: rol
        required: true
        schema:
          $ref: '#/definitions/schemas.Rol'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.Rol'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Actualizar rol
      tags:
      - Roles
  /api/v1/usuarios/login:
    post:
      consumes:
      - application/json
      description: Permite a un usuario autenticarse y obtener un token JWT. Solo
        usuarios aprobados pueden iniciar sesión.
      parameters:
      - description: Correo y contraseña
        in: body
        name: credentials
        required: true
        schema:
          properties:
            correo:
              type: string
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Token JWT generado
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Error de parsing del cuerpo
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Usuario no autorizado o contraseña incorrecta
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login de usuario
      tags:
      - Usuarios
  /api/v1/usuarios/solicitar:
    post:
      consumes:
      - application/json
      description: Permite a un usuario enviar una solicitud de registro. La contraseña
        se encripta y el usuario queda pendiente de aprobación.
      parameters:
      - description: Datos del usuario a registrar
        in: body
        name: usuario
        required: true
        schema:
          $ref: '#/definitions/schemas.Usuario'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Error de parsing del cuerpo
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Solicitar registro de usuario
      tags:
      - Usuarios
swagger: "2.0"